+++ bin/analyzer.js
